# REST-Web-Services/api-spec.yaml
openapi: 3.0.0
info:
  title: Posts API
  version: 1.0.0
  description: REST API for managing posts using Node.js, Express, and MySQL.

servers:
  - url: /api

paths:
  /posts:
    get:
      summary: List posts with pagination, search, and sorting.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Maximum number of posts to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of posts to skip.
        - in: query
          name: q
          schema:
            type: string
          description: Search string (matches topic or data).
        - in: query
          name: sort
          schema:
            type: string
            enum: [created_desc, created_asc]
            default: created_desc
          description: Sort order.
      responses:
        "200":
          description: List of posts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      limit: { type: integer }
                      offset: { type: integer }
                      count: { type: integer }
                      total: { type: integer }
                      sort: { type: string }
                      q: { type: string }
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create a new post.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic, data]
              properties:
                topic:
                  type: string
                data:
                  type: string
      responses:
        "201":
          description: Post created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /posts/{id}:
    get:
      summary: Fetch a single post by ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Invalid ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Post not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Replace an existing post.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic, data]
              properties:
                topic:
                  type: string
                data:
                  type: string
      responses:
        "200":
          description: Post updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Validation error or invalid ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Post not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Partially update a post.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                topic:
                  type: string
                data:
                  type: string
      responses:
        "200":
          description: Post updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Validation error or invalid ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Post not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a post.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Post deleted (no content).
        "400":
          description: Invalid ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Post not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /health:
    get:
      summary: Health check.
      responses:
        "200":
          description: Server is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

components:
  schemas:
    Post:
      type: object
      properties:
        id: { type: integer }
        topic: { type: string }
        data: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }